backport wildcard support
One of the main issues moving to a newer revision is that classes complain about
a newer bitbake and a newer bitbake is needed to add support for wildcards. In
an attempt to break this loop this backports wildcard support to an older bitbake:

Most notably:
31bc9af9c cooker: add support for using % as a wildcard in bbappend filename
991cbeedb cooker: Fix support for wildcards in bbappend filenames

The collection member is moved in this the patch though, since it not in
another object.